src/requirements.txt
asgiref==3.5.2
Django==4.0.5
Pillow==9.1.1
python-dotenv==0.20.0
sqlparse==0.4.2
psycopg2-binary==2.9.3
pytest==7.2.0
pytest-django==4.5.2
django-countries==7.3.2
authlib==1.0.1
requests==2.27.1
django-allauth==0.42.0
django-rosetta==0.9.8
djangorestframework==3.13.1
django-modeltranslation==0.18.4
django-crispy-forms==1.14.0
djangorestframework-simplejwt==5.2.0
django-social-share==2.3.0
python-dotenv==0.20.0
django-filter==22.1
django-widget-tweaks==1.4.12
python-decouple==3.6
django-autotranslate==1.2.0
redis==4.3.4
celery==5.2.3
django-celery-beat==2.3.0
django-cors-headers==3.13.0
drf-spectacular==0.24.2